// mobile/lib/email/types.ts
export type EmailStatus =
  | 'DRAFT'
  | 'QUEUED'
  | 'SCHEDULED'
  | 'SENT'
  | 'DELIVERED'
  | 'DELIVERY_DELAYED'
  | 'OPENED'
  | 'CLICKED'
  | 'BOUNCED'
  | 'COMPLAINED'
  | 'FAILED'
  | 'CANCELED'
  | 'RECEIVED';

export type EmailEventType =
  | 'QUEUED'
  | 'SCHEDULED'
  | 'SENT'
  | 'DELIVERED'
  | 'DELIVERY_DELAYED'
  | 'OPENED'
  | 'CLICKED'
  | 'BOUNCED'
  | 'COMPLAINED'
  | 'FAILED'
  | 'CANCELED'
  | 'RECEIVED'
  | 'REPLIED';
// ...
export interface EmailEvent {
  id: string;
  emailId: string;
  type: EmailEventType;
  payload?: unknown;
  link: string | null;
  occurredAt: string;
}